@@ -1,6 +1,6 @@
 {
   "name": "react-typeform",
-  "version": "0.1.1",
+  "version": "1.0.0",
   "description": "Fullscreen form or typeform in React",
   "main": "build/index.js",
   "keywords": [
@@ -11,37 +11,36 @@
     "react"
   ],
   "scripts": {
-   "test": "./node_modules/.bin/mocha --compilers js:babel-core/register --recursive",
-   "test-cov": "./node_modules/.bin/babel-node ./node_modules/.bin/isparta cover --report html --report text node_modules/.bin/_mocha",
-   "build": "rm -rf build && ./node_modules/.bin/babel src --out-dir build",
-   "clean": "rm -rf build",
-   "lint": "eslint src test"
- },
- "dependencies": {},
- "devDependencies": {
-   "babel-cli": "^6.10.1",
-   "babel-core": "^6.9.1",
-   "babel-eslint": "^6.0.4",
-   "babel-plugin-transform-react-jsx": "^6.8.0",
-   "babel-preset-es2015": "^6.9.0",
-   "babel-preset-stage-0": "^6.5.0",
-   "chai": "^3.5.0",
-   "enzyme": "^2.3.0",
-   "eslint": "^2.12.0",
-   "eslint-config-airbnb": "^9.0.1",
-   "eslint-plugin-import": "^1.8.1",
-   "eslint-plugin-jsx-a11y": "^1.4.2",
-   "eslint-plugin-react": "^5.1.1",
-   "isparta": "^4.0.0",
-   "mocha": "^2.5.3",
-   "react": "^15.1.0",
-   "react-dom": "^15.1.0"
- },
- "peerDependencies": {
-   "react": ">= ^0.14.3",
-   "react-dom": ">= ^0.14.3"
- },
- "repository": {
+    "test": "./node_modules/.bin/mocha --compilers js:babel-core/register --recursive",
+    "test-cov": "./node_modules/.bin/babel-node ./node_modules/.bin/isparta cover --report html --report text node_modules/.bin/_mocha",
+    "build": "rm -rf build && ./node_modules/.bin/babel src --out-dir build",
+    "clean": "rm -rf build",
+    "lint": "eslint src test"
+  },
+  "dependencies": {
+    "react": "^16.2.0",
+    "react-dom": "^15.1.0",
+    "proptypes": "^1.1.0"
+  },
+  "devDependencies": {
+    "babel-cli": "^6.26.0",
+    "babel-core": "^6.9.1",
+    "babel-eslint": "^6.0.4",
+    "babel-plugin-transform-react-jsx": "^6.8.0",
+    "babel-preset-env": "^1.6.1",
+    "babel-preset-es2015": "^6.24.1",
+    "babel-preset-stage-0": "^6.5.0",
+    "chai": "^3.5.0",
+    "enzyme": "^2.3.0",
+    "eslint": "^2.12.0",
+    "eslint-config-airbnb": "^9.0.1",
+    "eslint-plugin-import": "^1.8.1",
+    "eslint-plugin-jsx-a11y": "^1.4.2",
+    "eslint-plugin-react": "^5.1.1",
+    "isparta": "^4.0.0",
+    "mocha": "^2.5.3"
+  },
+  "repository": {
     "type": "git",
     "url": "git+https://github.com/mohithg/react-typeform.git"
   },
